explorind_project/requirements.txt
Django==1.5.1
South==0.8.1
dj-database-url==0.2.1
django-allauth==0.10.1
django-mptt==0.5.5
oauthlib==0.4.1
psycopg2==2.4.5
python-openid==2.2.5
requests==1.2.3
requests-oauthlib==0.3.2
